pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "playwright-mcp-recording"
version = "0.1.0"
description = "Playwright MCP Recording - Browser automation server with recording capabilities for Model Context Protocol"
readme = "README.md"
license = "MIT"
authors = [
    { name = "Playwright MCP Recording", email = "example@example.com" },
]
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=2.10.4",
    "playwright>=1.40.0",
    "playwright-stealth>=1.0.6",
    "pydantic>=2.11.7",
    "python-dotenv>=1.0.0",
]

[project.scripts]
playwright-mcp-recording = "src.server:cli"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.black]
line-length = 88
target-version = ['py310']

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]